+++ Packages/Accelib/Module/Accelib.Editor.AutoBuild/Architecture/AppConfig.cs
@@ -1,5 +1,5 @@
-﻿using System.Collections.Generic;
-using NaughtyAttributes;
+using System.Collections.Generic;
+using Sirenix.OdinInspector;
 using UnityEngine;
 
 namespace Accelib.Editor.Architecture
@@ -7,12 +7,14 @@ namespace Accelib.Editor.Architecture
     [System.Serializable]
     public class AppConfig
     {
-        [Header("앱 정보")]
+        [TitleGroup("앱 정보")]
         public string name;
+        [TitleGroup("앱 정보")]
         public string appID;
+        [TitleGroup("앱 정보")]
         public string liveBranch;
 
-        [Header("디포 정보")]
+        [TitleGroup("디포 정보")]
         public List<DepotConfig> depots;
     }
 }