+++ Packages/Accelib/Module/Accelib.Editor.AutoBuild/Architecture/EAddressablesBuildMode.cs
@@ -1,0 +1,10 @@
+namespace Accelib.Editor.Architecture
+{
+    /// <summary>어드레서블 빌드 모드</summary>
+    public enum EAddressablesBuildMode
+    {
+        Skip = 0,
+        ContentUpdate = 1,
+        CleanBuild = 2
+    }
+}